Add files via upload
The real-world validation dataset consists of 9,169 sensor measurements collected from real-world, in-field AC units. This "ground truth" was established by applying the same expert-defined rule-based logic to the raw sensor readings. This distinction is critical, the "sim-to-real" gap we are addressing is not a difference in labeling logic, but purely a difference in the feature distributions between the "clean" synthetic sensor readings and the "noisy" real-world sensor readings. This dataset is limited to the three conditions observed in the field, NORMAL (5,598 samples), MAINTENANCE 1 (2,412 samples), and MAINTENANCE 2 (1,159 samples). Furthermore, this dataset presents a significant test of model generalization. This real-world samples were collected from only 2-unit types (0.5 PK Inverter and 0.5 PK Non-Inverter), a small subset of the 12-unit types presents in the synthetic training data. This design forces the model to learn the general, underlying principles of fault behavior from the diverse 12-unit training set, rather than simply overfitting to the specific rules of the two target units. This dataset is held out and used exclusively by the validator component to score the fitness of each evolved model.